Split out converted attributes
