+++ src/exporters/datadog/transform/cloud.rs
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::exporters::datadog::transform::attributes::{ConvertedAttrMap, ConvertedAttrValue};
 use crate::semconv;
 use opentelemetry_semantic_conventions::attribute;
 use std::str::FromStr;
+use crate::exporters::datadog::transform::cv_attributes::{ConvertedAttrMap, ConvertedAttrValue};
 
 #[allow(dead_code)]
 const EC2_TAG_PREFIX: &str = "ec2.tag.";
