+++ src/exporters/datadog/transform/source.rs
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::exporters::datadog::transform::attributes::ConvertedAttrMap;
 use crate::exporters::datadog::transform::cloud;
 use crate::exporters::datadog::transform::k8s;
 use crate::semconv;
 use opentelemetry_semantic_conventions::attribute;
 use std::fmt::{Display, Formatter};
+use crate::exporters::datadog::transform::cv_attributes::ConvertedAttrMap;
 
 #[derive(Clone, Debug)]
 pub(crate) struct Source {
